+++ src/main/java/com/wayble/server/direction/repository/transportation/WheelchairInfoRepository.java
@@ -1,4 +1,4 @@
-package com.wayble.server.direction.repository;
+package com.wayble.server.direction.repository.transportation;
 
 import com.wayble.server.direction.entity.transportation.Wheelchair;
 import org.springframework.data.jpa.repository.JpaRepository;
